avance

+++ SIVEDI.Clases/TABLAS/EstadoActividadOfertas.cs
@@ -8,7 +8,7 @@ namespace SIVEDI.Clases.TABLAS
 {
     public class EstadoActividadOfertas
     {
-        public int CODIGO { get; set; }
+        public decimal CODIGO { get; set; }
         public int CODIGO_ESTADO { get; set; }
         public string NOMBRE { get; set; }
     }
